requirements.txt
asgiref==3.8.1
Django==4.2
django-jazzmin==2.6.0
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
openpyxl==3.1.5
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.5.3
xlsxwriter
